Move GET /size to an inseucre router, so no auth is needed
